requirements.txt
numpy
scipy
scikit-learn
Pillow
scikit-image>=0.19
pandas
pyarrow
matplotlib
seaborn
plotly
kaleido
jupyterlab
papermill
tensorboard
pytorch-lightning
python-dotenv
monai
GPUtil
psutil
segmentation-models-pytorch
kornia>=0.6.5
rich
threadpoolctl
commentjson
appdirs
blosc
torch
lazy-imports
